refactor socket_array

+++ include/swoole_api.h
@@ -33,9 +33,9 @@ SW_API swTimer_node* swoole_timer_get(long timer_id);
 SW_API uchar swoole_timer_clear(long timer_id);
 
 SW_API int swoole_event_init();
-SW_API uchar swoole_event_add(int fd, int events, int fdtype);
-SW_API uchar swoole_event_set(int fd, int events, int fdtype);
-SW_API uchar swoole_event_del(int fd);
+SW_API int swoole_event_add(int fd, int events, int fdtype);
+SW_API int swoole_event_set(int fd, int events, int fdtype);
+SW_API int swoole_event_del(int fd);
 SW_API void swoole_event_defer(swCallback cb, void *private_data);
 SW_API int swoole_event_write(int fd, const void *data, size_t len);
 SW_API int swoole_event_wait();
